% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\docType{data}
\name{agile_configuration}
\alias{agile_configuration}
\title{Agile Calendar Configuration}
\format{
The YAML (file) with the configuration and markers comprises
three sections: \code{configuration}, \code{agile_events}, and \code{markers} of
which the latter two have a somewhat interchangeable role.

\code{configuration} should have the following items below it (without
deeper hierarchies).

\describe{
\item{\code{project}}{A simple label to mark the name of a project or
something else that might be relevant to display. Currently
used more as mnemonic that actually used in the calendar.}
\item{\code{year}}{Simple number that is used to prefix the descriptive
labels for the increments and iterations (e.g. PI 22.x.y and
22 here then is the value for \code{year}.}
\item{\code{year_start}}{Date of the first Monday of the (next) Agile
year. This value is converted to a date stamp that is used to
offset all other calculations for laying out the Agile year
and calendar. Should be presented as YYYY-mm-dd.}
\item{\code{increment_length}, \code{iteration_length}}{The duration in
weeks of an increment (a quarter) and an iteration (a sprint).}
\item{\code{ip_length}}{The duration of the Innovation and Planning
iteration (in weeks).}
}

The \code{agile_events} section is used to add markers on the Agile
calendar that directly relate to the Agile cadence. This contrasts
(a little) with the \code{markers} section that is/could be used for
markers for (more personal) reminders about, for instance,
deadlines. Both sections comprise any number of elements which
then have one or more dates where to add markers. The name of
each subsection is used for the marker label.
}
\description{
The Agile calendar is defined and constructed using the
values inside the configuration YAML. A personal copy can be made
using the \code{\link[=copy_default_config]{copy_default_config()}} function.
}
