% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{cadence_markers}
\alias{cadence_markers}
\title{Mark Agile Cadence Related Events}
\usage{
cadence_markers(cal)
}
\value{
A list of ggplot2 geoms for drawing markers and adding labels
using ggrepel.
}
\description{
Add markers to the calendar for a set of predefined
markers related to events in the agile cadence.
}
