% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{copy_default_config}
\alias{copy_default_config}
\title{Copy Default Configuration}
\usage{
copy_default_config(path)
}
\arguments{
\item{path}{Path where to store the copy.}
}
\value{
Invisibly, checking whether copying was successful.
}
\description{
Make a copy of the default configuration to a location
outside of the package.
}
