% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{create_marker_dfr}
\alias{create_marker_dfr}
\title{Create Marker Data Frame / Tibble}
\usage{
create_marker_dfr(ev, ev_name, ...)
}
\arguments{
\item{ev}{One event from a list of events}

\item{ev_name}{Name of \code{ev} in the top level object.}

\item{...}{Further parameters to use in creating tibble.}
}
\value{
A one-row tibble with either a dstamp column for an absolute
date or, for instance, an \code{increment_wk} column for a relative
date.
}
\description{
The YAML configuration enables specify an absolute date as place for
a marker as well as using a relative date/location, for instance,
referring to the \emph{n}'th week in the increment. This function takes
care of the processing.
}
