% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_calendar.R
\name{generate_calendar}
\alias{generate_calendar}
\title{Generate an Agile Calendar}
\usage{
generate_calendar(cfg)
}
\arguments{
\item{cfg}{Configuration settings from a YAML project configuration.}
}
\value{
A data frame / tibble with a row for each week and variables
useful for visualizing the calendar.

\description{
\item{\code{dstamp}, \code{calendar_wk}}{Date of the Monday of a particular
week and the week number that belongs to that week (based on the
ISO standard).}
\item{\code{agile_wk}, \code{increment_wk}, \code{iteration_wk}}{Running week
number of, respectively, the Agile year, increment, and iteration.}
\item{\code{increment}, \code{iteration}}{Text labels for increments and
iterations.}
\item{\code{event}, \code{type}}{Markers to be mapped to weeks or dates in the
Agile calendar. Two types exist: markers related to Agile events
and rituals and (more personal) markers to, for instance, not
forget important deadlines.}}
}
\description{
Based on the \code{configuration} item in the YAML configuration
this function generates a dataframe that can be used to visualize
a calendar with default and additional markers from the remainder
of the configuration file.
}
