% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{is_nested}
\alias{is_nested}
\title{Check If List Is Nested}
\usage{
is_nested(l)
}
\arguments{
\item{l}{A list.}
}
\value{
Boolean whether \code{l} has nested lists.
}
\description{
Check If List Is Nested
}
