% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{parse_agile_events}
\alias{parse_agile_events}
\title{Parse Agile Events Overview}
\usage{
parse_agile_events(event_list)
}
\arguments{
\item{event_list}{Part of a structure from parsing the configuration
file using \code{\link[yaml:yaml.load]{yaml::yaml.load_file()}}.}
}
\description{
The agile events are part of the configuration file that
is parsed using the r-yaml package. Because (eventually) the agile
events are to be calculated relative to the agile cadence, this
separate function is used.
}
