% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Plot Agile Cadence Calendar}
\usage{
plot_calendar(cal)
}
\arguments{
\item{cal}{Dataframe with an agile calendar generated by the
respective function.}
}
\value{
A ggplot object
}
\description{
Plots an agile cadence calendar generated by
\code{\link[=generate_calendar]{generate_calendar()}} with a YAML configuration.
}
