% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_markers}
\alias{plot_markers}
\title{Mark Agile Cadence Related Events}
\usage{
plot_markers(cal, markers)
}
\arguments{
\item{cal}{An agile calendar tibble generated by \code{\link[=generate_calenar]{generate_calenar()}}.}

\item{markers}{String to indicate which markers to draw. Should be one
of "agile_events" or "markers".}
}
\value{
A list of ggplot2 geoms for drawing markers and adding labels
using ggrepel.
}
\description{
Add markers to the calendar for a set of predefined
markers related to events in the agile cadence.
}
