% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{read_config}
\alias{read_config}
\title{Load Configuration File}
\usage{
read_config(cfgfile)
}
\arguments{
\item{cfgfile}{A path to a configuration file.}
}
\description{
Load the configuration file for the generator.
}
