% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{today_marker}
\alias{today_marker}
\title{Draw Marker For Today}
\usage{
today_marker()
}
\value{
A list with a ggplot2 geom_vline geom that can be added to an
existing ggplot object.
}
\description{
Adds a marker to indicate today's date in the agile
calender visualization.
}
