% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_calendar.R
\name{generate_calendar}
\alias{generate_calendar}
\title{Generate an Agile Calendar}
\usage{
generate_calendar(cfg)
}
\arguments{
\item{cfg}{Configuration settings from a YAML project configuration.}
}
\value{
A data frame / tibble with a row for each week and variables
useful for visualizing the calendar.
}
\description{
Based on the \code{configuration} item in the YAML configuration
this function generates a dataframe that can be used to visualize
a calendar with default and additional markers from the remainder
of the configuration file.
}
